[PROJ-1816] Keep waitlist contracts backend-CI safe
